rename SetRegset to SetRegsetFromString
And updated related helpers similarly.

The "from string" versions of register set/get are intended
for the Remote Serial Protocol, and as such could be moved
into a subclass. Any such refactoring is for later, but this
patch adds consistency (the getters have FromString in the name),
and makes the separation from the normal get/set routines clearer.

Change-Id: I2ebbcb1a0e04420e441ff86b39a5a8daca018c76